+++ src/web-app/App.tsx
@@ -11,6 +11,8 @@ import { ContainerEngine, type OperatingSystem, type Program } from "@/env/Types
 import { DEFAULT_THEME } from "@/web-app/App.config";
 import "@/web-app/App.css";
 import "@/web-app/App.i18n";
+import { CURRENT_ENVIRONMENT } from "@/web-app/Environment";
+import { pathTo } from "@/web-app/Navigator";
 import AppErrorBoundary from "@/web-app/components/AppErrorBoundary";
 import { AppFooter } from "@/web-app/components/AppFooter";
 import { AppHeader } from "@/web-app/components/AppHeader";
@@ -18,8 +20,6 @@ import { AppLoading } from "@/web-app/components/AppLoading";
 import { AppSidebar } from "@/web-app/components/AppSidebar";
 import { StoreProvider } from "@/web-app/domain/store";
 import { AppBootstrapPhase, type AppStore, useStoreActions, useStoreState } from "@/web-app/domain/types";
-import { CURRENT_ENVIRONMENT } from "@/web-app/Environment";
-import { pathTo } from "@/web-app/Navigator";
 import { Screen as ContainerGenerateKubeScreen } from "@/web-app/screens/Container/GenerateKubeScreen";
 import { Screen as ContainerInspectScreen } from "@/web-app/screens/Container/InspectScreen";
 import { Screen as ContainerLogsScreen } from "@/web-app/screens/Container/LogsScreen";
@@ -44,6 +44,7 @@ import { Screen as PodProcessesScreen } from "@/web-app/screens/Pod/ProcessesScr
 import { Screen as RegistriesScreen } from "@/web-app/screens/Registry/ManageScreen";
 import { Screen as SecretInspectScreen } from "@/web-app/screens/Secret/InspectScreen";
 import { Screen as SecretsScreen } from "@/web-app/screens/Secret/ManageScreen";
+import { Screen as ConnectionInfoScreen } from "@/web-app/screens/Settings/ConnectionInfoScreen";
 import { Screen as SystemInfoScreen } from "@/web-app/screens/Settings/SystemInfoScreen";
 import { Screen as UserSettingsScreen } from "@/web-app/screens/Settings/UserSettingsScreen";
 import { Screen as TroubleshootScreen } from "@/web-app/screens/Troubleshoot/Troubleshoot";
@@ -78,6 +79,7 @@ const Screens = [
   VolumesScreen,
   VolumeInspectScreen,
   UserSettingsScreen,
+  ConnectionInfoScreen,
   SystemInfoScreen,
   TroubleshootScreen,
 ];